requirements.txt
streamlit>=1.32.0
mysql-connector-python>=8.3.0
pandas>=2.0.0
plotly>=5.18.0
PyPDF2>=3.0.0
reportlab>=4.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
